// src/app/shared/backend-api/model/setModel.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SetModel { 
    name?: string;
    code?: string;
    scryfallId?: string;
    type?: string;
    url?: string;
    releaseDate?: string;
    cards?: number;
    cardsReady?: number;
    iconUrl?: string;
    released?: boolean;
}